feat: add href to sidebar items
